pyproject.toml
[project]
name = "multi-intent-identification-app"
version = "0.1.1"
description = "HR Intent Identification Agent with Human-in-the-Loop via UiPath Action Center"
authors = [
    { name = "T MOHAMED Yaser", email = "1ammar.yaser@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-anthropic>=1.0.0",
    "uipath-langchain>=0.0.149",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.6.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["."]
